Aggiornamento dei percorsi dei moduli di configurazione e delle azioni di scambio
